Add dependencies to the spotifyd systemd service
Spotifyd crashes if it starts before either the sound card is configured or the network is available. Whilst this is not too bad, it causes a restart of the service when started at boot. It is nicer to just depend on the targets that systemd provides to signal availability of both.