Improve english on cache_path explanation
Define `the cache path' and remove ambiguity, 

Remove the repetitive 'is set to A so A is used'.

`have to exist' is not english. The bracketed text can also be removed because an explanation that the path must exist follows soon after.

First explain missing path and then shell expansion.

Remove redundant 'this means that ...'.

As a followup, what would happen if a relative path is provided? Is it allowed or must the path be absolute?