rewrite mainloop using async / await
The mainloop code is simplified by relying on the async / await syntax
instead of implementing all the futures ourselves. In doing so, this
also works, if a different user connects during the playback.

Additionally, the application chooses now between connecting using the
users credentials and starting discovery. If there are valid
credentials, no discovery stream is started at all, which prevents
buggy and surprising behaviour.